+++ input-stream/src/main/java/software/amazon/s3/dataaccelerator/io/logical/impl/DefaultLogicalIOImpl.java
@@ -17,23 +17,40 @@ package software.amazon.s3.dataaccelerator.io.logical.impl;
 
 import java.io.IOException;
 import lombok.NonNull;
+import software.amazon.s3.dataaccelerator.common.telemetry.Operation;
+import software.amazon.s3.dataaccelerator.common.telemetry.Telemetry;
+import software.amazon.s3.dataaccelerator.common.telemetry.TelemetryLevel;
 import software.amazon.s3.dataaccelerator.io.logical.LogicalIO;
 import software.amazon.s3.dataaccelerator.io.physical.PhysicalIO;
 import software.amazon.s3.dataaccelerator.request.ObjectMetadata;
+import software.amazon.s3.dataaccelerator.util.S3URI;
+import software.amazon.s3.dataaccelerator.util.StreamAttributes;
 
 /** The default implementation of a LogicalIO layer. Will be used for all non-parquet files. */
 public class DefaultLogicalIOImpl implements LogicalIO {
 
+  private static final String OPERATION_LOGICAL_READ = "logical.read";
+
   // Dependencies
+  private final S3URI s3URI;
   private final PhysicalIO physicalIO;
+  private final Telemetry telemetry;
+
+  // When is the LogicalIO instance created?
+  private final long birthTimestamp = System.nanoTime();
 
   /**
    * Constructs an instance of LogicalIOImpl.
    *
+   * @param s3URI the S3 URI of the object fetched
    * @param physicalIO underlying physical IO that knows how to fetch bytes
+   * @param telemetry an instance of telemetry
    */
-  public DefaultLogicalIOImpl(@NonNull PhysicalIO physicalIO) {
+  public DefaultLogicalIOImpl(
+      @NonNull S3URI s3URI, @NonNull PhysicalIO physicalIO, @NonNull Telemetry telemetry) {
+    this.s3URI = s3URI;
     this.physicalIO = physicalIO;
+    this.telemetry = telemetry;
   }
 
   /**
@@ -61,12 +78,36 @@ public class DefaultLogicalIOImpl implements LogicalIO {
   @Override
   public int read(byte[] buf, int off, int len, long position) throws IOException {
     // Perform read
-    return physicalIO.read(buf, off, len, position);
+    return telemetry.measureConditionally(
+        TelemetryLevel.VERBOSE,
+        () ->
+            Operation.builder()
+                .name(OPERATION_LOGICAL_READ)
+                .attribute(StreamAttributes.logicalReadPosition(position))
+                .attribute(StreamAttributes.logicalReadLength(len))
+                .attribute(StreamAttributes.uri(s3URI))
+                .attribute(
+                    StreamAttributes.logicalIORelativeTimestamp(System.nanoTime() - birthTimestamp))
+                .build(),
+        () -> physicalIO.read(buf, off, len, position),
+        bytesRead -> bytesRead > 1);
   }
 
   @Override
   public int readTail(byte[] buf, int off, int len) throws IOException {
-    return physicalIO.readTail(buf, off, len);
+    long contentLength = metadata().getContentLength();
+    long startOfRead = Math.max(0, contentLength - len);
+
+    return telemetry.measureVerbose(
+        () ->
+            Operation.builder()
+                .name(OPERATION_LOGICAL_READ)
+                .attribute(StreamAttributes.logicalReadPosition(startOfRead))
+                .attribute(StreamAttributes.uri(s3URI))
+                .attribute(
+                    StreamAttributes.logicalIORelativeTimestamp(System.nanoTime() - birthTimestamp))
+                .build(),
+        () -> physicalIO.readTail(buf, off, len));
   }
 
   /**