+++ object-client/src/main/java/software/amazon/s3/dataaccelerator/S3SdkObjectClient.java
@@ -161,6 +161,7 @@ public class S3SdkObjectClient implements ObjectClient {
             Operation.builder()
                 .name(ObjectClientTelemetry.OPERATION_GET)
                 .attribute(ObjectClientTelemetry.uri(getRequest.getS3Uri()))
+                .attribute(ObjectClientTelemetry.rangeLength(getRequest.getRange().getLength()))
                 .build(),
         s3AsyncClient
             .getObject(builder.build(), AsyncResponseTransformer.toBlockingInputStream())